3.4.0a - Slowly getting rid of Blender
This is a major partial release featuring:
- The node-canvas module

When I started this project, my intention was to use node-canvas to
render statistics, but compiling via node-gyp never worked, so I
switched to Blender, which has always been kind of a pain. Now there's a
prebuilt canvas package, so I had to use that now. It's a bit more code,
but like 100x faster and much much much very more flexible.

This is just a nightly build commit; The code is still not yet copied to
the RaspPi.